pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "documentationllm"
version = "0.1.0"
description = "Sistema de documentação baseado em LLM"
requires-python = ">=3.9"
authors = [
    {name = "Equipe de Desenvolvimento"}
]
dependencies = [
    "openai>=1.0.0",
    "python-dotenv>=1.0.0",
    "gitpython>=3.1.30",
    "pyyaml>=6.0",
    "markdown>=3.4.0",
    "beautifulsoup4>=4.11.0",
    "tqdm>=4.64.0",
    "colorama>=0.4.6",
    "click>=8.1.3",
    "rich>=13.0.0",
]

[project.scripts]
docllm = "documentationllm.cli:cli"

[tool.black]
line-length = 88
target-version = ['py39']
include = '\.pyi?$'

[tool.isort]
profile = "black"
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
line_length = 88

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -q --cov=src"
testpaths = [
    "tests",
]